gui_requirements.txt
# GUI Application Requirements
# ===========================

# Core GUI and Image Processing
tkinter  # Usually included with Python
Pillow>=9.0.0
opencv-python>=4.5.0

# Advanced Image Processing
scikit-image>=0.19.0
scipy>=1.7.0
PyWavelets>=1.3.0

# Machine Learning
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Additional utilities
pathlib
threading  # Usually included with Python
